+++ site/src/examples/Rtl/index.tsx
@@ -9,7 +9,6 @@ const App = () => {
   const [editorState, setEditorState] = useState(() => EditorState.fromJSON({
     text,
     ranges: [],
-    entityMap: {}
   }))
 
   return (