+++ site/src/examples/TimeTravel/index.tsx
@@ -10,7 +10,6 @@ const App = () => {
   const [editorState, setEditorState] = useState(() => EditorState.fromJSON({
     text,
     ranges: [],
-    entityMap: {}
   }))
 
   return (